@@ -4,7 +4,7 @@
   "private": true,
   "devDependencies": {
     "typings": "^1.0.4",
-    "typescript": "^1.8.10",
+    "typescript": "^2.0.0-beta",
 
     "gulp": "^3.9.1",
     "gulp-sass": "^2.3.1",
