@@ -1,10 +1,11 @@
 ﻿import { provideRouter, RouterConfig } from '@angular/router';
 
-import { BuildingList }    from './../building/building.list';
+import { GameList }    from './../building/game.list';
 import { Building }  from './../building/building';
 
 const routes: RouterConfig = [
-  { path: 'buildings', component: BuildingList },
+  { path: '', component: GameList },
+  { path: 'buildings', component: GameList },
   { path: 'building/:session', component: Building }
 ];
 
