@@ -2,43 +2,48 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using eSims.Data.Repository;
 using Microsoft.AspNetCore.Mvc;
 
 namespace eSims.Controllers
 {
-    [Route("api/[controller]")]
-    public class BuildingController : Controller
+  [Route("api/[controller]")]
+  public class BuildingController : ESimsController
+  {
+    public BuildingController(IApplicationRepository applicationRepository) : base(applicationRepository)
     {
-        // GET api/values
-        [HttpGet]
-        public IEnumerable<string> Get()
-        {
-            return new string[] { "value1", "value2" };
-        }
-
-        // GET api/values/5
-        [HttpGet("{id}")]
-        public string Get(int id)
-        {
-            return "value";
-        }
-
-        // POST api/values
-        [HttpPost]
-        public void Post([FromBody]string value)
-        {
-        }
-
-        // PUT api/values/5
-        [HttpPut("{id}")]
-        public void Put(int id, [FromBody]string value)
-        {
-        }
-
-        // DELETE api/values/5
-        [HttpDelete("{id}")]
-        public void Delete(int id)
-        {
-        }
     }
+
+    // GET api/values
+    [HttpGet]
+    public IEnumerable<string> Get()
+    {
+      return new string[] { "value1", "value2" };
+    }
+
+    // GET api/values/5
+    [HttpGet("{id}")]
+    public string Get(int id)
+    {
+      return "value";
+    }
+
+    // POST api/values
+    [HttpPost]
+    public void Post([FromBody]string value)
+    {
+    }
+
+    // PUT api/values/5
+    [HttpPut("{id}")]
+    public void Put(int id, [FromBody]string value)
+    {
+    }
+
+    // DELETE api/values/5
+    [HttpDelete("{id}")]
+    public void Delete(int id)
+    {
+    }
+  }
 }
